Enable decoding of subsumed instructions
This adds support for generating the most precise version of a given
instructino in the case of subsumsed instructions.

This also fixes some bugs in the instruction opcode formats / encoding.

This has also disabled the LSL and ROL and two branchind instructions
because they are exactly equivalent to others (sadtly).