Fix for decoding absolute addresses
There was a problem decoding the absolute addresses for JMP and CALL
instructions.  The problem was that I had mistakenly assumed the k bits
in the opcode would form the least significant bits.  However,
apparently not!!  This present some problems with my basic decoder
generator strategy.  However, by adding a "transform" I was able to work
around it.