Improved Connection Window; fix ROMS.
The connection window is now improved so that it does a better job of
trying to choose a pin for a given peripheral connection.  In addition,
with the inclusion of dynamic configuration for an AVR, I had to update
the ROMs.  That's because they simple weren't connecting to the correct
register bits.